database/populate_letters/requirements.txt
praw>=7.7.0
supabase>=1.0.3
python-dotenv>=1.0.0
faker>=18.10.1 